+++ crates/opsml_server/tests/api/card.rs
@@ -315,9 +315,6 @@ async fn test_opsml_server_card_datacard_crud() {
     let response = helper.send_oneshot(request).await;
     assert_eq!(response.status(), StatusCode::OK);
 
-    // sleep for 1 sec
-    tokio::time::sleep(Duration::from_millis(100)).await;
-
     let body = response.into_body().collect().await.unwrap().to_bytes();
     let create_response: CreateCardResponse = serde_json::from_slice(&body).unwrap();
     assert!(create_response.registered);
@@ -450,9 +447,6 @@ async fn test_opsml_server_card_modelcard_crud() {
     let body = response.into_body().collect().await.unwrap().to_bytes();
     let create_response: CreateCardResponse = serde_json::from_slice(&body).unwrap();
     assert!(create_response.registered);
-
-    // sleep for 1 second to allow the key to be stored
-    tokio::time::sleep(Duration::from_millis(100)).await;
 
     let load_request = CardQueryArgs {
         uid: Some(create_response.key.uid.clone()),
@@ -740,9 +734,6 @@ async fn test_opsml_server_card_auditcard_crud() {
     let response = helper.send_oneshot(request).await;
     assert_eq!(response.status(), StatusCode::OK);
 
-    // sleep for 1 sec
-    tokio::time::sleep(Duration::from_millis(100)).await;
-
     let body = response.into_body().collect().await.unwrap().to_bytes();
     let create_response: CreateCardResponse = serde_json::from_slice(&body).unwrap();
     assert!(create_response.registered);
@@ -878,9 +869,6 @@ async fn test_opsml_server_card_get_card() {
     let body = response.into_body().collect().await.unwrap().to_bytes();
     let create_response: CreateCardResponse = serde_json::from_slice(&body).unwrap();
 
-    // wait 1 sec
-    tokio::time::sleep(Duration::from_millis(100)).await;
-
     create_card_metadata(create_response.key.clone());
     //
     //// 2. Now test getting the card
@@ -966,10 +954,11 @@ async fn test_opsml_server_card_get_readme() {
 
     let response = helper.send_oneshot(request).await;
     assert_eq!(response.status(), StatusCode::OK);
+
     //
     let body = response.into_body().collect().await.unwrap().to_bytes();
     let card_readme: ReadeMe = serde_json::from_slice(&body).unwrap();
-
+    assert_eq!(card_readme.exists, true);
     assert_eq!(card_readme.readme, "This is a test README");
 
     //
