+++ crates/opsml_server/tests/api/mod.rs
@@ -1,2 +1,5 @@
 pub mod card;
+pub mod experiment;
+pub mod files;
 pub mod login;
+pub mod user;
